Major FrontEnd Impovments

+++ filmbox-app/src/main/java/com/janchabik/filmbox/repo/CreatorRepository.java
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @RepositoryRestResource(path = "creators", collectionResourceRel = "creators")
 @CrossOrigin(origins = "http://localhost:4200")
-public interface CreatorRepository extends JpaRepository<Creator,Integer> {
+public interface CreatorRepository extends JpaRepository<Creator,Long> {
 }
