+++ filmbox-app/src/main/java/com/janchabik/filmbox/repo/FilmRepository.java
@@ -7,6 +7,6 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @RepositoryRestResource(path = "films", collectionResourceRel = "films")
 @CrossOrigin(origins = "http://localhost:4200")
-public interface FilmRepository extends JpaRepository<Film,Integer> {
+public interface FilmRepository extends JpaRepository<Film,Long> {
 
 }
